+++ OHAttributedLabel/Source/OHTouchesGestureRecognizer.m
@@ -49,7 +49,7 @@
     CGPoint currentPoint = [touch locationInView:self.view];
     CGFloat distanceX = (currentPoint.x - _startPoint.x);
     CGFloat distanceY = (currentPoint.y - _startPoint.y);
-    CGFloat distance = sqrtf(distanceX * distanceX + distanceY * distanceY);
+    CGFloat distance = sqrt(distanceX * distanceX + distanceY * distanceY);
     if (distance > 10.0f) {
         self.state = UIGestureRecognizerStateCancelled;
     } else {
